Don't do int cmoves for IEEE comparisons.
Protect int cmove from raising an assertion if it is trying to do an int
conditional move where the test involves floating point comparisons that
can't easily be reversed due to NaNs.

The code used to generate the condition, and possibly reverse the condition if
ISEL could not handle it by rewriting the OP in the comparison rtx.
Unfortunately there are some conditions like UNLE that can't easily be reversed
due to NaNs.

The patch changes this so that it does the reversal before generating the
comparison.  If the comparison cannot be reversed, it just returns false,
saying that we can't do an int conditional move in this case.

2022-02-17  Michael Meissner  <meissner@the-meissners.org>

gcc/
	PR target/104256
	* config/rs6000/rs6000.cc (rs6000_emit_int_cmove): Don't do
	integer conditional moves if the test needs to be reversed and
	there isn't a direct reverse comparison.

gcc/testsuite/
	PR target/104256
	* gcc.target/powerpc/ppc-fortran/pr104254.f90: New test.