rs6000: Reset -mpower8-fusion for power9 inlining power8 functions [PR102059]
If a function is specifically declared as a power8 function, you can't
inline in functions that are specified with power9 or power10 options.
The issue is -mpower8-fusion is cleared when you use -mcpu=power9 or
-mcpu=power10.  Since -mpower8-fusion does not change which instructions
we can generate, the solution here is to ignore that flag when determing
whether we can inline a function or not.

2022-02-18   Michael Meissner  <meissner@linux.ibm.com>

gcc/
	PR target/102059
	* config/rs6000/rs6000.cc (rs6000_can_inline_p): Don't test for
	power8-fusion if the caller was power9 or power10.

gcc/testsuite/
	PR target/102059
	* gcc.target/powerpc/pr102059-4.c: New test.

(Backport of a submitted trunk patch that isn't upstream yet.)