fold-const: Fix optimization of COND_EXPR with capabilities
fold_ternary_loc tries to optimize A ? 0 : 1 into !A. To do this, it
negates the condition and then tries to fold_convert_loc the condition
into the result type, which may be a capability type. It is of course
invalid to try and just convert a non-capability type into a capability
type. Instead, in this case, we need to null-derive a capability using
REPLACE_ADDRESS_VALUE.

gcc/ChangeLog:

	* fold-const.c (fold_ternary_loc): Fix A ? 0 : 1 -> !A
	optimization for capabilities.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/morello/fold-ternary.c: New test.