Temporary step to get multilib using a bit better
This is a hack for the moment.  The approach we should use is to create
a canonicalised version of the architecture string for choosing
multilibs (similar to commit f58d03b5d did for arm).

What we're doing here is artificially handling the particular
architecture combination that glibc uses to build specially by linking
that to the standard C64 multilib.

N.b. we also need to update the `genmultilib` script.  This script
turned a description of names into a pathname.  However, because we
wanted to include a multilib option without it forming a directory path
(we did not want the -march=morello+c64 option to be included as part of
the path) a temporary in this script had multiple `/`s at the end of
a directory and only removed one of them.

The reason we only want `purecap` as a directory name is that the
`purecap` part defines the ABI while the other architecture options are
not the determining factor (we can have purecap with many different
architecture options).

This change removes all trailing `/`s, which stops a break in the glibc
build scripts when using multilibs.