Optimize signed DImode -> TImode on power10.
On power10, GCC tries to optimize the signed conversion from DImode to
TImode by using the vextsd2q instruction.  However to generate this
instruction, it would have to generate 3 direct moves (1 from the GPR
registers to the altivec registers, and 2 from the altivec registers to
the GPR register).

This patch adds code back in to use the shift right immediate instruction
to do the conversion if the target/source is GPR registers.

2022-02-25   Michael Meissner  <meissner@linux.ibm.com>

gcc/
	* config/rs6000/vsx.md (mtvsrdd_diti_w1): Delete.
	(extendditi2): Replace with code to deal with both GPR registers
	and with altivec registers.