Avoid defining MULTILIB_REUSE rule when no purecap multilib
This MULTILIB_REUSE rule is there as a temporary measure to ensure that
if someone uses the arguments `-mabi=purecap -march=morello+c64+nolse`
that links to the same multilib as `-mabi=purecap -march=morello+c64`.

This is a temporary step, and we plan to make a proper fix following the
same mechanism as was used for CDE arguments in the arm backend.

Unconditionally using the MULTILIB_REUSE variable breaks building when
the purecap multilib is not being built.  This is because the
genmultilib script requires all multilib reuse targets to exist.

This patch puts a condition on defining the variable so it is not
defined when we are not building a purecap multilib.