Fix some morello tests for stock AArch64
known-outside-bounds.c was checking for a warning that we emit when
writing out a capability that will end up out of bounds.  This warning
will not occur on fake-capability since it's based on the machinery of
writing capabilities out.

packed-structures.c was emitting a large amount of warnings on unaligned
pointers.  We just remove those warnings since the whole point of the
test was about this precise case.

unwinding.c (and no-frame-chain-unwind.c which used the same code) had
two compilation problems.  The first was that `uintptr_t` was not
defined but was used.  Thes econd was that we had an array of function
pointers which was statically initialised.  For dynamically linked
programs these function addresses are not known at link time.  Here we
initialise the function pointers dynamically to account for this.