Rework gen_lowpart_if_possible, fix ICE
For the testcase added in this patch, expand_debug_expr would create an
rtx of the form:

(truncate:DI (mem:CADI (symbol_ref:CADI ...)))

and we would later ICE in rtlanal.c:num_sign_bit_copies1, where we
(quite reasonably) don't expect to see an rtx of this form. Since the
TRUNCATE was originally created through the simplify_* routines, this
should be simplified to just:

(mem:DI (symbol_ref:CADI ...))

Specifically, simplify_unary_operation_1 in the TRUNCATE case has logic
to attempt using a SUBREG instead, where this is possible. It does this
using gen_lowpart_if_possible. The MEM_P case of gen_lowpart_if_possible
looks like it should perform the desired transformation, but
unfortunately it checks memory_address_addr_space_p on the resulting MEM
address, and bails if this doesn't hold. In this case,
memory_address_addr_space_p never held for the MEM in the first place,
and in general need not hold for all MEMs in arbitrary rtl.

This patch reworks gen_lowpart_if_possible to use lowpart_subreg if the
initial gen_lowpart_common call fails and the rtx has non-VOIDmode.
With this change, we can now perform the desired transformation.

gcc/ChangeLog:

	* rtlhooks.c (gen_lowpart_if_possible): Rework to use
	lowpart_subreg.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/morello/debug-truncate-ice.c: New test.