dwarf: create BTF decl and type tag DIEs
The "btf_decl_tag" and "btf_type_tag" attributes are handled by
constructing DW_TAG_LLVM_annotation DIEs. The DIEs are children of the
declarations or types which they annotate, and convey the annotation via
a string constant.

Currently, all generation of these DIEs is gated behind
btf_debuginfo_p (). That is, they will not be generated nor output
unless BTF debug information is generated. The DIEs will be output in
DWARF if both -gbtf and -gdwarf are supplied by the user.

gcc/

	* dwarf2out.cc (gen_btf_decl_tag_dies): New function.
	(gen_btf_type_tag_dies): Likewise.
	(modified_type_die): Call them here, if appropriate.
	(gen_formal_parameter_die): Likewise.
	(gen_typedef_die): Likewise.
	(gen_type_die): Likewise.
	(gen_decl_die): Likewise.