ctfc: Add support to pass through BTF annotations
BTF generation currently relies on the internal CTF representation to
convert debug info from DWARF dies. This patch adds a new internal
header, "ctf-int.h", which defines CTF kinds to be used internally to
represent BTF tags which must pass through the CTF container. It also
adds a new type for representing information specific to those tags, and
a member for that type in ctf_dtdef.

This patch also updates ctf_add_reftype to accept a const char * name,
and add it for the newly added type.

gcc/

	* ctf-int.h: New file.
	* ctfc.cc (ctf_add_reftype): Add NAME parameter. Pass it to
	ctf_add_generic call.
	(ctf_add_pointer): Update ctf_add_reftype call accordingly.
	* ctfc.h (ctf_add_reftype): Analogous change.
	(ctf_btf_annotation): New.
	(ctf_dtdef): Add member for it.
	(enum ctf_dtu_d_union_enum): Likewise.
	* dwarf2ctf.cc (gen_ctf_modifier_type): Update call to
	ctf_add_reftype accordingly.