Keep the cheri_capability attribute
This patch keeps the cheri_capability attribute in TYPE_ATTRIBUTES
and uses it to determine the mode of pointers.  In some ways it
feels a bit nasty to be calling layout_type in the middle of
build_type_attribute_qual_variant, but I think it's the right
interface.  If the starting type passed to the function is not
a capability type then it should not (on hybrid targets) have
a capability mode, so the caller can't change the mode first.
Changing the mode later would leave a significant window in which
the type is internally inconsistent.