morello: Improve handling of REPLACE_ADDRESS_VALUE constants
The initial problem here was an ICE in
varasm.c:tree_innermost_capability via
varasm.c:initializer_constant_valid_p_1. The code in these functions
only expected to see an INTEGER_CST as the base capability in a
REPLACE_ADDRESS_VALUE. However, we can of course derive an __intcap from
a pointer, simply by casting the pointer to an __intcap (and optionally
adding an offset), as in the testcase added with this patch.

We adapt the code in the above functions to handle this case.

After making this change, it was found that the backend couldn't handle
assembling the resulting rtxes on pure-capability Morello. This can be
remedied by introducing a new RTL simplification rule to transform:

(replace_address_value:M C (plus:OM C.CV X)) --> (pointer_plus:M C X).

where M is a capability mode, OM is its offset mode, C is a capability, C.CV is
its address value, and X is an arbitrary rtx.

The backend can of course already handle the POINTER_PLUS form.

Note that this bug was originally triggered by a C++ testcase,
g++.dg/cpp0x/pr88410.C. However, the underlying issue can also be triggered from
C, as with the testcase in this patch.

gcc/ChangeLog:

	* simplify-rtx.c (lowparts_p): If the simplified rtxes are not
	registers, return true if they compare equal by pointer
	equality. This handles e.g. subregs of SYMBOL_REFs.
	(simplify_binary_operation_1): New rule to simplify
	REPLACE_ADDRESS_VALUE to POINTER_PLUS where possible.
	(test_replace_address_value_simplifications): New test.
	* varasm.c (tree_innermost_capability): Also handle ADDR_EXPR as
	the innermost capability.
	(initializer_constant_valid_p_1): Handle constants derived from
	ADDR_EXPRs.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/morello/intcap-pointer-plus-ice.c: New test.