More libgcc type fixes
In execute_stack_op our `result` is defined as an _Unwind_CapWord since
it needs to produce a capability on capability architectures, and on
non-capability architectures it must produce some standard sized value.

When the operation this function was performing was a DW_OP_abs, we
check if the value is negative and turn it positive if needs be.

The check for if the value is negative needed a signed value, and
originally we used _Unwind_Sword.  However this ended up differently
sized as well as differently signed for some non-capability architectures
(e.g. ilp32), and that meant the comparison would always return true.

aarch64-unwind.h was passing a capability as the PC address to look for
to _Unwind_Find_FDE but we now want just an address.

aarch64_demangle_return_addr didn't mark its _Unwind_FrameState argument
as unused.

The C personality function using `read_encoded_value` pointing at an
unsigned value when that function takes a pointer to a capability (so
would have overwritten some things on the stack).

Similar was happening in `execute_cfa_program` where we read into an
_Unwind_Address pointer rather than an _Unwind_Ptr one.

We change a type in an expression inside `execute_stack_op` to avoid a
provenance warning (even though the default provenance choice was good
for us).