Rename _Unwind_CapWord to _Unwind_Word
This means that the API external applications saw has not changed.
We have always called these arguments _Unwind_Word, and we maintain
that.

To summarise the changes in types that we've made in the unwinder for
Morello:
 - _Unwind_Word and _Unwind_Ptr have both been split into two different
   types.  One for capabilities and one for non-capabilities.
 - _Unwind_Word and _Unwind_Ptr now refer to the capability forms, while
   _Unwind_WordAddr and _Unwind_Address have been introduced to refer to
   the non-capability forms.
 - On non-capability systems the capability and non-capability forms are
   the same as each other.
(similar change for _Unwind_Word has been made for its signed version
_Unwind_Sword).

The reason we need the distinction between _Unwind_Word and _Unwind_Ptr
is that the _Unwind_Ptr type represents a pointer at the language level,
while an _Unwind_Word represents "whatever type the backend decided to
use for the unwinder".  Often (e.g. in the case of AArch64 ilp32) this
is the equivalent of GCC's Pmode rather than ptr_mode (i.e. the type
that represents an address at the machine level rather than the type
that represents an address at the language level).

In this patch we assume that pointers for purecap targets are of the
same mode as the Pmode equivalent for the purecap architecture.  This
seems to be a safe assumption since otherwise how would a purecap pointer
be implemented.

Another patch will adjust the meaning of `unwind_word` inside the
compiler to ensure that it aligns with the meaning we now give it in the
unwinder.