Revert previous change of _Unwind_Find_FDE declaration
We had previously changed the declaration of _Unwind_Find_FDE to take an
_Unwind_Address rather than a void *.  While this makes it clear that
the function acts on addresses rather than pointers it also changes a
relatively widely used function throughout the unwinder.

While the _Unwind_Find_FDE function is not declared by any headers that
get provided to the user of libgcc, it is still exported in the libgcc
dynamic library.  Hence this is still an ABI determining function.

We also fix some whitespace changes that have been introduced against
trunk to make things a little tidier.

N.b. in this patch we do a few superficially useless casts (turn a
sealed pointer into an address, then cast it to a pointer).  This is
because the logical meaning of `__builtin_code_address_from_pointer` is
to return the address from a pointer, and even though the integer value
is the same it's still nice to maintain that abstraction level.
It also means that it's clear the value we're providing to
_Unwind_Find_FDE is not something that can be dereferenced.