Convert GCC-internal unwind_word to scalar_addr_mode
Now we've switched the definition of _Unwind_Word in the unwinder to
represent a capability on capability architectures, it makes sense to
change the `unwind_word` mode in GCC itself to mean a capability mode on
capability architectures.

The propagation of this means:
 - targetm.unwind_word_mode and targetm.eh_return_filter_mode have their
   return type changed to scalar_addr_mode.  Similarly their default
   implementations have the same change.
 - We add a new implementation of this hook for AArch64 that returns
   Pmode.
 - In order to implement the _Unwind_WordAddr type in libgcc, we
   introduce a new unwind_word_addr mode attribute that indicates the
   offset_mode of unwind_word.
 - This new unwind_word_addr mode attribute is then used in libgcc.

If we want to define the `unwind_word` mode as a capability mode when
that's what's needed for `_Unwind_Word`, then disallowing the use of
`attribute((mode(unwind_word)))` on integers (which was originally done
in commit cfc5512ea6ccb) breaks the main motivating case for this mode
definition.

This patch also allows the use of capability modes in a mode attribute
on integers to define an INTCAP_TYPE.  This does allow a slightly
confusing syntax of defining an INTCAP_TYPE based on a modified unsigned
integer, but it means the definition of `unwind_word` and `_Unwind_Word`
is consistent throughout the compiler and libgcc.

That had the extra benefit that we no longer used __UINTPTR_TYPE__ for
_Unwind_Ptr.  The uintptr_t value is guaranteed to be able to hold any
value of a `void *`, but it's not guaranteed to be the same size as a
`void *`.

The existing implementation (before Morello changes) defined _Unwind_Ptr
with a typedef using `__attribute__((__mode__(__pointer__)))` rather
than with `__UINTPTR_TYPE__`.  This *is* guaranteed to be the same size
as a `void *`.

Since the unwinder reads in pointers from DWARF information, and this
information is ABI defined to have the same size as a `void *`, we can
not guarantee that changing to `__UINTPTR_TYPE__` will always work -- it
may read in extra data from the dwarf information that would end up
having a different value than the `void *` stored there.

Being able to revert to `__attribute__((__mode__(__pointer__)))` avoids
the above problem.

During implementation we now ensure that `build_intcap_type_for_mode`
caches the types that it can return (similar to how
`build_pointer_type_for_mode` also caches the types that it can return).

We needed that behaviour for this change because we use
`lang_hooks.types.type_for_mode` in `handle_mode_attribute` to set the
type node which we use after the application of the relevant mode
attribute.

As it stands, `lang_hooks.types.type_for_mode` uses
`build_intcap_type_for_mode`, and since that function builds a new type
each time we end up with differing types representing the same thing.
Those differing types cause problems because they compare as different
and hence type-handling code does the wrong thing on them.

In this particular case we saw an ICE trying to find the common type for
an expression because the types were not comparing equal and we had no
way to handle two INTCAP_TYPE's with different `TYPE_MAIN_VARIANT`
types.  Not having such a way makes sense because there are not
`INTCAP_TYPE`'s with differing `TYPE_MAIN_VARIANT`'s in any capability
architecture as yet.

N.b. Because we now return a capability from `__builtin_extend_pointer`
the testcase `aarch64/asm-4.c` now works as it did for non-capability.