morello: Fix up aarch64_output_mi_thunk
This patch fixes up aarch64_output_mi_thunk for capabilities. Firstly,
in the case with multiple (non-virtual) inheritance, where the
vcall_offset is zero, we make sure to pass Pmode temporaries to
aarch64_add_offset. This function requires the temporaries to be in the
same mode that is passed in. Note that, if at any point an offset_mode
temporary is required, aarch64_add_offset can just call drop_capability
to obtain the (hard) register in the appropriate mode.

For the case with virtual inheritance, where vcall_offset is nonzero, we
change the last step of the sequence. As it stands, the code calls
gen_add2_insn to offset this_rtx, effectively computing `this_rtx =
this_rtx + p_temp1`, but p_temp1 is a pointer, not an offset, so we ICE.
The value we load into p_temp1 is really just an offset which is stored
in the vtable as a pointer, so we change the code to load it in POmode
instead.

This fixes ICEs in the C++ testsuite with -mfake-capability where
multiple inheritance thunks are required.