In fake-capability, do not offset fake-capability DWARF register numbers
A simple fix to fix an ICE that I previously introduced
to fake-capability bootstrap and regression testing.

The change here is to consider TARGET_CAPABILITY_FAKE as well
as the machine mode when deciding if to offset a reg-no
by 198. This is to exclude fake-cap from applying the
offset (in fake-cap we could still get here with the mode
as CADImode, so the offset would incorrectly be applied)