Make power8-fusion depend only on tuning for power8.
The power8 fusion support used to be set automatically when -mcpu=power8 or
-mtune=power8 was used, and it was cleared for other cpu's.  However, if you
used the target attribute or target #pragma to change the default cpu type or
tuning, you would get an error that a target specifiction option mismatch
occurred.

This code removes the -mpower8-fusion and -mpower8-fusion-sign options, and
only enables power8 fusion if we are tuning for a power8.  Power8 sign fusion
is only enabled if we are tuning for a power8 and we have -O3 option or
higher.

2022-03-09   Michael Meissner  <meissner@linux.ibm.com>

gcc/
	PR target/102059
	* config/rs6000/rs6000-cpus.def (OTHER_FUSION_MASKS): Delete.
	(ISA_3_0_MASKS_SERVER): Don't clear the fusion masks.
	(POWERPC_MASKS): Remove OPTION_MASK_P8_FUSION.
	* config/rs6000/rs6000.cc (rs6000_option_override_internal):
	Delete code that set the power8 fusion options automatically.
	(rs6000_opt_masks): Allow #pragma target and attribute target to set
	power8-fusion and power8-fusion-sign, but these no longer represent
	options that the user can set.
	* config/rs6000/rs6000.h (TARGET_P8_FUSION): Set automaticaly based on
	whether we are tuning for a power8.
	(TARGET_P8_FUSION_SIGN): Likewise.
	(MASK_P8_FUSION): Delete.
	* config/rs6000/rs6000.opt (-mpower8-fusion): Recognize the option but
	ignore the no form and warn that the option was removed for the regular
	form.
	(-mpower8-fusion-sign): Likewise.
	* doc/invoke.texi (RS/6000 and PowerPC Options): Delete -mpower8-fusion
	and -mpower8-fusion-sign.

gcc/testsuite/
	PR target/102059
	* gcc.target/powerpc/pr102059-4.c: New test.