Introduce mem{cpy,move,cmp}_c builtins
GCC uses the mem{cpy,move,cmp} library functions as a fallback for block
memory operations that we don't want to open-code.

In purecap this is just fine since the functions take capabilities and
all pointers are capabilities.

For Hybrid, these functions take non-capability pointers.  This means
that if we decide to fall back to these library functions when asked to
do these operations on capability pointers we would throw away the
security behaviours on said capability pointers.

To account for this, hybrid capability targets provide versions of the
above functions which take capability pointer arguments.

Strictly speaking, it is optional for these functions to be provided by
the runtime (as documented in the "hybrid C guide", but this is the
only reasonable fallback for a compiler and LLVM already relies on these
functions existing rather than open-code arbitrary block memory
operations).
https://github.com/CTSRD_CHERI/cheri_hybrid_c_guide/

This patch updates GCC to use the capability enabled versions of these
memory operations as needs be (i.e. when we want to perform these
operations on memory pointed to by a capability pointer).

We do this by introducing three new builtins accordingly.  These
builtins are only defined when we are targetting a hybrid capability
target.

The builtins are used automatically when one of the
`emit_block_{copy,move,comp}_via_libcall` functions is used on a
hybrid capability target with at least one capability-pointer argument.

Notes:
 1) In using these functions we define a builtin version for each one.
    Such builtins are not in Morello LLVM.  They are added here simply
    because of the authors aesthetic tastes -- having the builtins
    defined means that the use of `memcpy` and `memcpy_c` is more
    similar both to the user and in GCC source code.  In order to avoid
    defining the builtin we could use `init_one_libfunc` and
    `emit_library_call`.
 2) In order to define these new builtins, we needed new types to use.
    The new types needed to be capabilities.  We only define this
    builtins for a hybrid capability target.  There is no existing
    mechanism to only define a type in builtin-types.def for a given
    target.  We gave the types a fallback of the non-capability version.
    Overall this means that on a non-capability target we have defined
    function types that have a name referring to capabilities but do not
    use capabilities.  Even though these functions are not used on a
    non-capability target this is slightly ugly.
    It seemed easier and less-intrusive than adding a conditional to all
    the function definition macros in each of the backends.
 3) We define new `global_trees` nodes `cap_const_ptr_type_node` and
    `cap_ptr_type_node`.  These are just added for convenience.