morello: Fix regression with -Wcheri-provenance
In commit 6404652ab0fc9904b48224d4a8a2e955f30c4e5e, we move the definition of
the -Wcheri-provenance option from c-family/c.opt to common.opt, but in the
process we drop the Var() part of the definition. This leads to an ICE in
opts.c:common_handle_option. The option-handling code requires that either there
is some explicit code to handle an option, or it has a Var() which can be
assigned to, which is the default behaviour.

This patch restores the Var() part of the -Wcheri-provenance definition which
was lost in the commit mentioned above.

gcc/ChangeLog:

	* common.opt (Wcheri-provenance): Restore Var().

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/morello/wno-cheri-provenance.c: New test.