Rename get_address_mode to mem_address_mode
I wanted to add a helper function that returned the mode of
an address rtx (only necessary because of modeless CONST_INTs).
The obvious name was get_address_mode, but that name was
already taken by a function that returns the mode of a
MEM's address.

This patch renames the existing function to mem_address_mode,
which hopefully makes the operation a little clearer.  The new
get_address_mode will take two arguments, so new uses of the old
get_address_mode would be caught/rejected at build time.