Avoid trans-capability MEM_REF relaxation in SLSR
There is an optimisation in SLSR that takes a COMPONENT_REF (or similar)
whose base is a MEM_REF, and it splits that out into a MEM_REF using a
different offset (pre-calculated using known constants).

When performing this optimisation it uses the type of the offset for
different things, and ends up changing the type of the MEM_REF base to
the type of the offset.
This is problematic, since if the base is a capability type and the
offset is not this is throwing away the capability features.

Here we just disable this optimisation if the two types are not
compatible.