cp: Fix folding of REPLACE_ADDRESS_VALUE
For code such as (char *)1 in the C++ frontend, we were producing the
following IR:

.REPLACE_ADDRESS_VALUE(0B, 1)

but this should be folded to a ponter-typed INTEGER_CST with value 1.
This was causing various problems including e.g. an ICE at expand time
when we see the IFN in a context where we don't expect it.

The reason for the lack of folding is that when the C++ frontend sees
(char *)1, it wraps the INTEGER_CST 1 in a NON_LVALUE_EXPR. Prior to
this change, fold_build_replace_address_value_loc didn't know to look
inside such a tree code, so failed to do the expected folding.

To remedy this, we teach fold_build_replace_address_value_loc to strip
location wrappers on its operands (like various other fold routines),
and adjust cp_cap_from_noncap to ensure we preserve any location
wrappers that might get stripped in order to do the folding. This
follows existing practice in other conversion routines in the C++
frontend (see e.g. convert_to_pointer_maybe_fold).

gcc/cp/ChangeLog:

	* cvt.c (cp_cap_from_noncap): Call preserve_any_location_wrapper
	to restore location wrappers that might get stripped from
	constant folding.

gcc/ChangeLog:

	* fold-const.c (fold_build_replace_address_value_loc): Call
	STRIP_ANY_LOCATION_WRAPPER on the operands.